Separate ESP32 Azure IoT Kit sample into main and core modules
